Fix an overall bit-reading behavior at the very end of buffer.
It has been permitted that `j40__always_refill` can be called with
the *less* number of requested bits than absolutely needed. This
behavior is forbidden in `j40__refill` macro, but Huffman decoding
needed an access to possibly more bits than actual codeword, so in this
case `j40__always_refill` will just try to read as many bits as
possible, which can be less than the requested size (`max_len`).

It turns out that there are three different bugs in this logic.
The first is a possible infinite loop, because `j40__refill_buffer` can
fail to refill a single byte. The second is that at the very end of file
or section, it may (rarely) fail to read as many bytes as possible, so
the very last codeword might be incorrectly chopped off. Conversely,
the third is that `j40__prefix_code` failed to check if the last
codeword is complete so it might accept an incorrect codestream. This
commit fixes all three bugs.