Fix an incorrect assertion in j40__cluster_map.
The intention was to ensure that seen[0..256) matches 1+ 0*, where there
are exactly `*num_clusters` copies of 1s. Given non-zero `num_dist` (now
properly asserted, but it was always the case), there should be at least
one seen[k] which is 1, but the assertion incorrectly checked for the
number of leading 1s which can be 0 for some incorrect codestream where
k > 0. The assertion has now delayed so that it first matches 1* 0* and
counts the number of 1s at the very end.