Initial implementation of j40__limits.
This is required for efficient fuzzing. (More stringent limitations are
planned in the future when j40_limit API function is available.) Every
limit in the Annex M is checked, except for `icc_output_size` which
requires a big change to j40__icc (still impacts fuzzing),
`num_splines`, `num_patches` and those related to frame durations.