Disable LeakSanitizer for the dj40-o0g target.
LeakSanitizer doesn't work with strace etc. enabled, which is essential
for gdb to function. As this target is mainly used to debug J40, we
disable LSAN unless requested in this case. We only do this when
J40_DEBUG is set, because ASAN should be also usable for the main target
and we would want to enable LSAN then.